Rename bgmTrack, clarify code a bit
Added some comments about the scene switcher, audio system, etc.

SceneTreeTween was backported to 3.6, should look into that later.

Reparent the Interactable Area2D of the PGO as the root. Maybe going forward, having it be the root of most objects should be the convention to make it easier to edit and organize.